Add files via upload
Refactoring, or rewriting, code is a vital skill for keeping codebases up to date. You don't want to rewrite code from scratch when you could just fix and improve something that's already there. In this project, you're going to refactor a Connect Four game to make it object-oriented! Through this project, you can see for yourself how much easier everything is when you use OOP.

 

For example, what if you wanted to take this code and make several games of Connect Four run simultaneously in the same window? Without refactoring this project, you'd have to write hundreds of lines of code to enable this. If you refactor it properly, you'll only have to write three or four lines of code to do this. 

 

Quick note, when using this, sometimes functionality that was previously working will break unexpectedly. Even though that's frustrating, you must power through to get the hang of the keyword. Eventually, it will make your life a lot easier.